detection param presa fini

+++ Lemin/srcs/libft/includes/libft.h
@@ -18,7 +18,6 @@
 # include <sys/types.h>
 # include <sys/stat.h>
 # include <fcntl.h>
-# include "libftprintf.h"
 
 void				ft_bzero(void *s, size_t n);
 void				*ft_memset(void *b, int c, size_t len);
@@ -42,7 +41,7 @@ char				*ft_strnstr(const char *s1, const char *s2, size_t n);
 int					ft_strcmp(const char *s1, const char *s2);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
 
-int					ft_atoi(const char *str);
+long int			ft_atoi(const char *str);
 int					ft_atoibase(char *nb, int base);
 
 int					ft_isalpha(int c);
